requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.5
django-js-asset==1.2.2
django-mptt==0.11.0
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2020.5
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
